Split createSessionTimelineStore into record/prune/query modules
- Map ベースの内部状態を TimelineState 型として明示化し、
  イベント記録（record.ts）、保持期間の枝刈り・closePane（prune.ts）、
  範囲別集計（query.ts）を state を第一引数に取る関数へ分離
- store.ts は TimelineState の保持と各モジュール関数の合成のみに縮小
  （公開 API・挙動は変更なし）
- 分離前に、記録タイムスタンプが逆順で届いた場合の closeAtMs/
  lastBoundaryMs クランプ挙動（record 2 パターン・closePane 1 パターン）
  を固定するテストを追加
