+++ apps/server/src/state-timeline/store.ts
@@ -1,398 +1,58 @@
-import type {
-  SessionStateTimeline,
-  SessionStateTimelineItem,
-  SessionStateTimelineRange,
-  SessionStateTimelineSource,
-  SessionStateValue,
-} from "@vde-monitor/shared";
-
+import { type ClosePaneInput, closePane as closePaneEvent, prunePane } from "./prune";
 import {
-  aggregateRepoTimelineSegments,
-  buildTimelineBoundaries,
-  clipTimelineEventToInterval,
-} from "./timeline-aggregation";
+  DEFAULT_RETENTION_MS,
+  type GetRepoTimelineInput,
+  type GetTimelineInput,
+  getTimeline as queryPaneTimeline,
+  getRepoTimeline as queryRepoTimeline,
+} from "./query";
+import { type RecordStateTransitionInput, recordStateTransition } from "./record";
 import {
   type SessionTimelinePersistedEvent,
   type SessionTimelinePersistedEvents,
   type TimelineEvent,
   normalizeRestoredPaneEvents,
-  toIso,
 } from "./timeline-restore";
-import { parseIsoToMs } from "../utils/time";
-
-const RANGE_MS: Record<SessionStateTimelineRange, number> = {
-  "15m": 15 * 60 * 1000,
-  "1h": 60 * 60 * 1000,
-  "3h": 3 * 60 * 60 * 1000,
-  "6h": 6 * 60 * 60 * 1000,
-  "24h": 24 * 60 * 60 * 1000,
-  "3d": 3 * 24 * 60 * 60 * 1000,
-  "7d": 7 * 24 * 60 * 60 * 1000,
-  "14d": 14 * 24 * 60 * 60 * 1000,
-  "30d": 30 * 24 * 60 * 60 * 1000,
-};
-
-const DEFAULT_RETENTION_MS = RANGE_MS["30d"];
-const MAX_TIMELINE_ITEMS = 10_000;
-const DEFAULT_LIMIT_BY_RANGE: Record<SessionStateTimelineRange, number> = {
-  "15m": 200,
-  "1h": 300,
-  "3h": 700,
-  "6h": 1_500,
-  "24h": 5_000,
-  "3d": 7_000,
-  "7d": 10_000,
-  "14d": 10_000,
-  "30d": 10_000,
-};
 
 type StoreOptions = {
   now?: () => Date;
   retentionMs?: number;
 };
 
-type RecordStateTransitionInput = {
-  paneId: string;
-  state: SessionStateValue;
-  reason: string;
-  repoRoot?: string | null;
-  at?: string;
-  source?: SessionStateTimelineSource;
+export type TimelineState = {
+  eventsByPane: Map<string, TimelineEvent[]>;
+  sequence: number;
+  now: () => Date;
+  retentionMs: number;
 };
-
-type ClosePaneInput = {
-  paneId: string;
-  at?: string;
-};
-
-type GetTimelineInput = {
-  paneId: string;
-  range?: SessionStateTimelineRange;
-  limit?: number;
-};
-
-type GetRepoTimelineInput = {
-  paneId: string;
-  paneIds: string[];
-  range?: SessionStateTimelineRange;
-  limit?: number;
-  aggregateReason?: string;
-  itemIdPrefix?: string;
-};
-
-const createEmptyTotals = (): Record<SessionStateValue, number> => ({
-  RUNNING: 0,
-  WAITING_INPUT: 0,
-  WAITING_PERMISSION: 0,
-  SHELL: 0,
-  UNKNOWN: 0,
-});
-
-const resolveTimelineLimit = (range: SessionStateTimelineRange, limit: number | undefined) => {
-  const fallback = DEFAULT_LIMIT_BY_RANGE[range];
-  if (limit == null || !Number.isFinite(limit)) {
-    return fallback;
-  }
-  return Math.min(MAX_TIMELINE_ITEMS, Math.max(1, Math.floor(limit)));
-};
-
-const TIMELINE_STATE_PRIORITY: SessionStateValue[] = [
-  "WAITING_PERMISSION",
-  "RUNNING",
-  "WAITING_INPUT",
-  "SHELL",
-  "UNKNOWN",
-];
 
 export const createSessionTimelineStore = (options: StoreOptions = {}) => {
-  const now = options.now ?? (() => new Date());
-  const retentionMs = options.retentionMs ?? DEFAULT_RETENTION_MS;
-  const eventsByPane = new Map<string, TimelineEvent[]>();
-  let sequence = 0;
-
-  const nextId = (paneId: string, startedAtMs: number) => {
-    sequence += 1;
-    return `${paneId}:${startedAtMs}:${sequence}`;
+  const timelineState: TimelineState = {
+    eventsByPane: new Map(),
+    sequence: 0,
+    now: options.now ?? (() => new Date()),
+    retentionMs: options.retentionMs ?? DEFAULT_RETENTION_MS,
   };
 
-  const resolveAtMs = (at: string | undefined, fallbackMs: number) => {
-    const parsed = parseIsoToMs(at);
-    return parsed == null ? fallbackMs : parsed;
-  };
+  const record = (input: RecordStateTransitionInput) => recordStateTransition(timelineState, input);
 
-  const getOrCreatePaneEvents = (paneId: string) => {
-    const existing = eventsByPane.get(paneId);
-    if (existing) {
-      return existing;
-    }
-    const next: TimelineEvent[] = [];
-    eventsByPane.set(paneId, next);
-    return next;
-  };
+  const closePane = (input: ClosePaneInput) => closePaneEvent(timelineState, input);
 
-  const prunePane = (paneId: string, nowMs: number) => {
-    const events = eventsByPane.get(paneId);
-    if (!events || events.length === 0) {
-      return;
-    }
-    const thresholdMs = nowMs - retentionMs;
-    const retained = events.filter((event) => {
-      if (!event.endedAt) {
-        return true;
-      }
-      const endedAtMs = parseIsoToMs(event.endedAt);
-      if (endedAtMs == null) {
-        return true;
-      }
-      return endedAtMs >= thresholdMs;
-    });
+  const getTimeline = (input: GetTimelineInput) => queryPaneTimeline(timelineState, input);
 
-    events.splice(0, events.length, ...retained);
-  };
-
-  const record = ({
-    paneId,
-    state,
-    reason,
-    repoRoot = null,
-    at,
-    source = "poll",
-  }: RecordStateTransitionInput) => {
-    if (!paneId) {
-      return;
-    }
-    const nowMs = now().getTime();
-    const events = getOrCreatePaneEvents(paneId);
-    prunePane(paneId, nowMs);
-
-    let atMs = resolveAtMs(at, nowMs);
-    const last = events.at(-1);
-    if (last) {
-      const lastStartMs = parseIsoToMs(last.startedAt) ?? atMs;
-      const lastBoundaryMs = parseIsoToMs(last.endedAt) ?? lastStartMs;
-      if (atMs < lastBoundaryMs) {
-        atMs = lastBoundaryMs;
-      }
-
-      if (!last.endedAt) {
-        if (
-          last.state === state &&
-          last.reason === reason &&
-          last.source === source &&
-          last.repoRoot === repoRoot
-        ) {
-          return;
-        }
-        const closeAtMs = Math.max(lastStartMs, atMs);
-        last.endedAt = toIso(closeAtMs);
-      }
-    }
-
-    events.push({
-      id: nextId(paneId, atMs),
-      paneId,
-      state,
-      reason,
-      repoRoot,
-      startedAt: toIso(atMs),
-      endedAt: null,
-      source,
-    });
-    prunePane(paneId, nowMs);
-  };
-
-  const closePane = ({ paneId, at }: ClosePaneInput) => {
-    if (!paneId) {
-      return;
-    }
-    const events = eventsByPane.get(paneId);
-    if (!events || events.length === 0) {
-      return;
-    }
-    const last = events.at(-1);
-    if (!last || last.endedAt) {
-      return;
-    }
-    const nowMs = now().getTime();
-    const startedAtMs = parseIsoToMs(last.startedAt) ?? nowMs;
-    const atMs = resolveAtMs(at, nowMs);
-    last.endedAt = toIso(Math.max(startedAtMs, atMs));
-    prunePane(paneId, nowMs);
-  };
-
-  const getTimeline = ({ paneId, range = "1h", limit }: GetTimelineInput): SessionStateTimeline => {
-    const nowMs = now().getTime();
-    const nowIso = toIso(nowMs);
-    prunePane(paneId, nowMs);
-
-    const rangeMs = RANGE_MS[range];
-    const rangeStartMs = nowMs - rangeMs;
-    const resolvedLimit = resolveTimelineLimit(range, limit);
-
-    const events = eventsByPane.get(paneId) ?? [];
-    const totals = createEmptyTotals();
-
-    const withDuration = events
-      .map<SessionStateTimelineItem | null>((event) => {
-        const interval = clipTimelineEventToInterval({
-          event,
-          rangeStartMs,
-          nowMs,
-          parseIso: parseIsoToMs,
-        });
-        if (!interval) {
-          return null;
-        }
-        const durationMs = Math.max(0, interval.endedAtMs - interval.startedAtMs);
-        totals[event.state] += durationMs;
-        return {
-          id: event.id,
-          paneId: event.paneId,
-          state: event.state,
-          reason: event.reason,
-          startedAt: event.startedAt,
-          endedAt: event.endedAt,
-          source: event.source,
-          durationMs,
-        };
-      })
-      .filter((event): event is SessionStateTimelineItem => event != null)
-      .sort((a, b) => {
-        const aMs = parseIsoToMs(a.startedAt) ?? 0;
-        const bMs = parseIsoToMs(b.startedAt) ?? 0;
-        return bMs - aMs;
-      });
-
-    const items = withDuration.slice(0, resolvedLimit);
-    const current = items.find((item) => item.endedAt == null) ?? null;
-    return {
-      paneId,
-      now: nowIso,
-      range,
-      items,
-      totalsMs: totals,
-      current,
-    };
-  };
-
-  const resolveDominantState = (states: SessionStateValue[]) => {
-    for (const state of TIMELINE_STATE_PRIORITY) {
-      if (states.includes(state)) {
-        return state;
-      }
-    }
-    return "UNKNOWN" as SessionStateValue;
-  };
-
-  const resolveDominantSource = (sources: SessionStateTimelineSource[]) => {
-    if (sources.includes("hook")) {
-      return "hook" as SessionStateTimelineSource;
-    }
-    if (sources.includes("restore")) {
-      return "restore" as SessionStateTimelineSource;
-    }
-    return "poll" as SessionStateTimelineSource;
-  };
-
-  const getRepoTimeline = ({
-    paneId,
-    paneIds,
-    range = "1h",
-    limit,
-    aggregateReason = "repo:aggregate",
-    itemIdPrefix = "repo",
-  }: GetRepoTimelineInput): SessionStateTimeline => {
-    const nowMs = now().getTime();
-    const nowIso = toIso(nowMs);
-    const rangeMs = RANGE_MS[range];
-    const rangeStartMs = nowMs - rangeMs;
-    const resolvedLimit = resolveTimelineLimit(range, limit);
-    const totals = createEmptyTotals();
-
-    const uniquePaneIds = Array.from(new Set(paneIds.filter(Boolean)));
-    uniquePaneIds.forEach((candidatePaneId) => {
-      prunePane(candidatePaneId, nowMs);
-    });
-
-    const intervals = uniquePaneIds.flatMap((candidatePaneId) => {
-      const events = eventsByPane.get(candidatePaneId) ?? [];
-      return events
-        .map((event) =>
-          clipTimelineEventToInterval({
-            event,
-            rangeStartMs,
-            nowMs,
-            parseIso: parseIsoToMs,
-          }),
-        )
-        .filter((item): item is NonNullable<typeof item> => item != null);
-    });
-
-    if (intervals.length === 0) {
-      return {
-        paneId,
-        now: nowIso,
-        range,
-        items: [],
-        totalsMs: totals,
-        current: null,
-      };
-    }
-
-    const boundaries = buildTimelineBoundaries({ rangeStartMs, nowMs, intervals });
-    const segments = aggregateRepoTimelineSegments({
-      intervals,
-      boundaries,
-      nowMs,
-      resolveDominantState,
-      resolveDominantSource,
-      aggregateReason,
-    });
-
-    const items = segments
-      .map<SessionStateTimelineItem>((segment, index) => {
-        const durationMs = Math.max(0, segment.endedAtMs - segment.startedAtMs);
-        totals[segment.state] += durationMs;
-        return {
-          id: `${itemIdPrefix}:${paneId}:${segment.startedAtMs}:${index}`,
-          paneId,
-          state: segment.state,
-          reason: segment.reason,
-          startedAt: toIso(segment.startedAtMs),
-          endedAt: segment.isOpen ? null : toIso(segment.endedAtMs),
-          durationMs,
-          source: segment.source,
-        };
-      })
-      .sort((left, right) => {
-        const leftMs = parseIsoToMs(left.startedAt) ?? 0;
-        const rightMs = parseIsoToMs(right.startedAt) ?? 0;
-        return rightMs - leftMs;
-      })
-      .slice(0, resolvedLimit);
-
-    const current = items.find((item) => item.endedAt == null) ?? null;
-    return {
-      paneId,
-      now: nowIso,
-      range,
-      items,
-      totalsMs: totals,
-      current,
-    };
-  };
+  const getRepoTimeline = (input: GetRepoTimelineInput) => queryRepoTimeline(timelineState, input);
 
   const reset = () => {
-    eventsByPane.clear();
+    timelineState.eventsByPane.clear();
   };
 
   const serialize = (): SessionTimelinePersistedEvents => {
-    const nowMs = now().getTime();
-    eventsByPane.forEach((_events, paneId) => {
-      prunePane(paneId, nowMs);
+    const nowMs = timelineState.now().getTime();
+    timelineState.eventsByPane.forEach((_events, paneId) => {
+      prunePane(timelineState, paneId, nowMs);
     });
     const output: SessionTimelinePersistedEvents = {};
-    eventsByPane.forEach((events, paneId) => {
+    timelineState.eventsByPane.forEach((events, paneId) => {
       if (events.length === 0) {
         return;
       }
@@ -408,8 +68,8 @@ export const createSessionTimelineStore = (options: StoreOptions = {}) => {
       | null
       | undefined,
   ) => {
-    eventsByPane.clear();
-    sequence = 0;
+    timelineState.eventsByPane.clear();
+    timelineState.sequence = 0;
     if (!persisted) {
       return;
     }
@@ -422,14 +82,14 @@ export const createSessionTimelineStore = (options: StoreOptions = {}) => {
       }
       const { events: restored, maxSequence } = normalizeRestoredPaneEvents(paneId, events);
       if (restored.length > 0) {
-        eventsByPane.set(paneId, restored);
+        timelineState.eventsByPane.set(paneId, restored);
       }
-      sequence = Math.max(sequence, maxSequence);
+      timelineState.sequence = Math.max(timelineState.sequence, maxSequence);
     });
 
-    const nowMs = now().getTime();
-    eventsByPane.forEach((_events, paneId) => {
-      prunePane(paneId, nowMs);
+    const nowMs = timelineState.now().getTime();
+    timelineState.eventsByPane.forEach((_events, paneId) => {
+      prunePane(timelineState, paneId, nowMs);
     });
   };
